refactor: move and rename DetailsCompleted and DownloadConfirmation
