--- src/components/DownloadAndShareComponents/DetailsCompleted/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./DetailsCompleted";
