+++ src/components/TeacherComponents/DownloadConfirmation/DownloadConfirmation.tsx
@@ -4,7 +4,7 @@ import Flex from "@/components/SharedComponents/Flex";
 import { Heading, P } from "@/components/SharedComponents/Typography";
 import Svg from "@/components/SharedComponents/Svg";
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
-import NextLessonContainer from "@/components/DownloadComponents/NextLessonContainer";
+import DownloadConfirmationNextLessonContainer from "@/components/TeacherComponents/DownloadConfirmationNextLessonContainer";
 import { NextLesson } from "@/node-lib/curriculum-api-2023/queries/lessonDownloads/lessonDownloads.schema";
 import { TrackFns } from "@/context/Analytics/AnalyticsProvider";
 
@@ -113,7 +113,7 @@ const DownloadConfirmation: FC<DownloadConfirmationProps> = ({
       </Flex>
 
       {displayNextLessonContainer && isNextLessonsAvailable && (
-        <NextLessonContainer
+        <DownloadConfirmationNextLessonContainer
           programmeSlug={programmeSlug}
           unitSlug={unitSlug}
           unitTitle={unitTitle}
