+++ src/components/TeacherComponents/ResourcePageDetailsCompleted/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ResourcePageDetailsCompleted";
